kafka_ex v0.5.0 KafkaEx.Protocol.Common

A collection of common request generation and response parsing functions for the Kafka wire protocol.

Functions

parse_topics(topics_size, arg2, mod)

topic_data(list)

Generate the wire representation for a list of topics.