kafka_ex v0.5.0 KafkaEx.Protocol.Produce.Message

key: is used for partition assignment, can be nil, when none is provided it is defaulted to nil

value: is the message to be written to kafka logs.

Types

t :: %KafkaEx.Protocol.Produce.Message{key: binary, value: binary}