kafka_ex v0.6.0 KafkaEx.Protocol.Produce.Request

require_acks: indicates how many acknowledgements the servers should receive before responding to the request. If it is 0 the server will not send any response (this is the only case where the server will not reply to a request). If it is 1, the server will wait the data is written to the local log before sending a response. If it is -1 the server will block until the message is committed by all in sync replicas before sending a response. For any number > 1 the server will block waiting for this number of acknowledgements to occur (but the server will never wait for more acknowledgements than there are in-sync replicas), default is 0

timeout: provides a maximum time in milliseconds the server can await the receipt of the number of acknowledgements in RequiredAcks, default is 100 milliseconds

Types

t :: %KafkaEx.Protocol.Produce.Request{compression: atom, messages: list, partition: integer, required_acks: integer, timeout: integer, topic: binary}